refactor(data): move Condition to filter package
The Condition class is more of a filter (it's a very thin layer above).
It might be refactored to become a union of Filter and Condition.
