--- xoai-data-provider/src/main/java/io/gdcc/xoai/dataprovider/model/conditions/package-info.java
@@ -1,1 +1,0 @@
-package io.gdcc.xoai.dataprovider.model.conditions;
